+++ NEWS.md
@@ -1,5 +1,8 @@
+# ezplot 0.7.0
+- added instantaneous lift to perf_df
+
 # ezplot 0.6.6
-- added quantiles option to perf-df
+- added quantiles option to perf_df
 - fixed precision-recall plots when categorical target is used
 
 # ezplot 0.6.5